Moved importModal to a different file
